Reorders parameters in seas() function to put the feature container first. (#8)
The `seas()` function is (really) polymorphic over the object that contains the feature data/ranks/etc, for instance: are you running GSEA over a ranked vector of score? A dataframe that contains features that need to be ranked? Perhaps an expression container w/ design and contrasts (like limma::camera or mroast(?)).

This means that we should really put the object that stores these things (`x`) first, instead of the GeneSetDb object.

The first three parameters of the `seas()` function are now:

1. `x`, "feature" representation (or expression container)
2. `gsd` the `GeneSetDb`
3. `methods` a character vector that enumerates the GSEA methods you want to run.

Closes #5