caclulateIndividualLogFC handles $genes data.frame with colnames that collide with DGE stats.
We were getting errors when a `*List` object with a `$genes`
data.frame that had columns with names that collided with the
statistics from the internal DGE analysis. For example if there
was a `$genes$pval` column.

This fix renames the offending `$genes` column names to
`$genes.feature`, like `$genes$pval.features`

Also, we removed a use of `within(df, { ... })` because that's just
too much.

Thanks for @sandersen12 for the bug report!