Removes the dispatch to lapply or bplapply in seas based on BPPARAM.
We used to use standard `lapply` in the call to `seas` when
`BPPARAM` was set to `BiocParallel::SerialParam` so that we could
ensure the RNG that was used was the same that was preset in the
unit tests via `set.seed()` so they would pass.

Now, instead of having that "code smell" in the package code, we rather
are forcing ourselves to write a bit more tortured unitests by foregoing
the use of `set.seed()` and writing test code that generates a result
that relies on some type of randomness / permutation (like fgsea) to
be called within a `BiocParallel::bplapply(1, function(i) { ... })`
construct.